Artist Tab UI Jitter, Featured on Albums, Multiple columns on larger screens (#330)
Add some springing to the artist banner artwork so that it doesn't jitter as much

Adds a tab for "Featured on" albums, where albums that the artist is featured on will go. This tab does not appear if there are no featured on albums

Adds multicolumn support for larger screens (tested on iPad Pro). Now there will be as many columns as the width allows for cards