A-Z Selector for Artists, Prep for the same for Albums, Style Fixes, Verbiage Adjustments (#373)
Lots of overall improvements to styling, color usage, and wording

Adds more iconography, adjust settings tab to hide "Labs", enabled via 5 taps on "Made with <3" message

Adds flashlist, for faster lists. Lots of work on the library to add an alphabetical selector for skipping to a particular spot in the list quickly. Will be iterated on as we build out more ways to sort and view your library!

adds additional dns fixes for connecting to a local domain