Library Improvements (#500)
Make gesture in the alphabetical selector better, especially when tapping on a letter. When users tap on a letter now, they will be brought to that section of the artists list

Fix issue where Artists in the Library would show "0 Albums". This will now instead display the artist's genres instead